added method getInitials from full name
