+++ .gitignore
@@ -3,3 +3,4 @@ allure-reports
 node_modules
 package-lock.json
 package.json
+axios.methods.js
