+++ helpers/methods.js
@@ -4,4 +4,12 @@ function clearInput(selector) {
         selector.keys('Delete');
     }
 }
-module.exports = { clearInput };            // curly braces because we can add some other methods in the arr
+
+function getInitials(name){
+    let initials = "";
+    for (let i = 0; i < name.length; i++){
+        if (name[i] === " ")
+            return (initials = name.charAt(name[0])+name.charAt(i + 1)).toUpperCase();
+    }
+}
+module.exports = { clearInput, getInitials };            // curly braces because we can add some other methods in the arr
