+++ test/specs/profileTest.spec.js
@@ -1,19 +1,27 @@
 const LoginPage = require('../pageobjects/Login.page');
 const ProfilePage = require('../pageobjects/Profile.page');
-const MenuPage = require("../pageobjects/Menu.page");
+const GlobalNavigation = require("../pageobjects/GlobalNavigation.page");
+const { getInitials } = require("../../helpers/methods")
 
 describe("Profile", () => {
 
     before(async () => {
         await browser.maximizeWindow();
-        await LoginPage.fillLoginCredentials('Manya111@test.com', 'Manya111@');
-        await LoginPage.clickLoginBtn();
     })
 
     it('Should redirect on Profile Page', async () => {
-        await MenuPage.clickMenu();
-        await MenuPage.clickProfile();
-        await expect(ProfilePage.title).toHaveText("user");
-        expect(ProfilePage.title).toBeExisting().true;
+        await LoginPage.fillLoginCredentials('Manya111@test.com', 'Manya111@');
+        await LoginPage.clickLoginBtn();
+        await GlobalNavigation.btnMenu.click();
+        await GlobalNavigation.profileOption.click();
+        const res = await ProfilePage.title.getText();
+        expect(res).toEqual("user");
     });
+
+    it("ImageLetter should match FullName", async () => {
+        const fullName = await ProfilePage.profileName.getText();
+        const nameInit  = await getInitials(fullName);
+        const imageInit = await ProfilePage.profileImageInitials.getText();
+        await expect(nameInit).toEqual(imageInit);
+    })
 });
