+++ wdio.conf.js
@@ -5,7 +5,7 @@ exports.config = {
     // ====================
     // Runner Configuration
     // ====================
-    //
+    //npm
     //
     // ==================
     // Specify Test Files
